main/requirements.txt
Keras==2.4.3
numpy>=1.19.5
pandas>=0.19
scikit-learn>=0.24.1
tensorflow==2.4.1
streamlit>=0.82.0
Pillow==7.2.0